pyproject.toml
[tool.poetry]
name = "es-statistics"
version = "0.1.0"
description = ""
authors = ["cucuridas <cucuridas@gamil.com>"]
readme = "README.md"
packages = [{include = "es_statistics"}]

[tool.poetry.dependencies]
python = "^3.11"
elasticsearch = "^8.6.2"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.4"
uvloop = "^0.17.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
